refactor(heal): split task.rs per heal kind (#6293)
